Split config file and add auth file support
