+++ pkg/config/app.go
@@ -11,6 +11,7 @@ const (
 	LogFile              = "core.log"
 	PidFile              = "core.pid"
 	CfgFile              = "config.toml"
+	AuthFile             = "auth.toml"
 	UserDir              = "user"
 	ApiRequestTimeout    = 30 * time.Second
 	SuccessSoundFilename = "success.wav"
