+++ pkg/config/config.go
@@ -2,12 +2,11 @@ package config
 
 import (
 	"errors"
-	"fmt"
+	"net/url"
 	"os"
 	"path/filepath"
 	"regexp"
 	"runtime"
-	"slices"
 	"strings"
 	"sync"
 
@@ -26,75 +25,21 @@ const (
 )
 
 type Values struct {
-	ConfigSchema int       `toml:"config_schema"`
-	DebugLogging bool      `toml:"debug_logging"`
-	Audio        Audio     `toml:"audio,omitempty"`
-	Readers      Readers   `toml:"readers,omitempty"`
-	Systems      Systems   `toml:"systems,omitempty"`
-	Launchers    Launchers `toml:"launchers,omitempty"`
-	ZapScript    ZapScript `toml:"zapscript,omitempty"`
-	Service      Service   `toml:"service,omitempty"`
-	Mappings     Mappings  `toml:"mappings,omitempty"`
-	Groovy       Groovy    `toml:"groovy,omitempty"`
+	ConfigSchema int                        `toml:"config_schema"`
+	DebugLogging bool                       `toml:"debug_logging"`
+	Audio        Audio                      `toml:"audio,omitempty"`
+	Readers      Readers                    `toml:"readers,omitempty"`
+	Systems      Systems                    `toml:"systems,omitempty"`
+	Launchers    Launchers                  `toml:"launchers,omitempty"`
+	ZapScript    ZapScript                  `toml:"zapscript,omitempty"`
+	Service      Service                    `toml:"service,omitempty"`
+	Mappings     Mappings                   `toml:"mappings,omitempty"`
+	Groovy       Groovy                     `toml:"groovy,omitempty"`
+	Auth         map[string]CredentialEntry `toml:"auth,omitempty"`
 }
 
 type Audio struct {
 	ScanFeedback bool `toml:"scan_feedback,omitempty"`
-}
-
-type Readers struct {
-	AutoDetect bool             `toml:"auto_detect"`
-	Scan       ReadersScan      `toml:"scan,omitempty"`
-	Connect    []ReadersConnect `toml:"connect,omitempty"`
-}
-
-type ReadersScan struct {
-	Mode         string   `toml:"mode"`
-	ExitDelay    float32  `toml:"exit_delay,omitempty"`
-	IgnoreSystem []string `toml:"ignore_system,omitempty"`
-	OnScan       string   `toml:"on_scan,omitempty"`
-	OnRemove     string   `toml:"on_remove,omitempty"`
-}
-
-type ReadersConnect struct {
-	Driver   string `toml:"driver"`
-	Path     string `toml:"path,omitempty"`
-	IDSource string `toml:"id_source,omitempty"`
-}
-
-func (r ReadersConnect) ConnectionString() string {
-	return fmt.Sprintf("%s:%s", r.Driver, r.Path)
-}
-
-type Systems struct {
-	Default []SystemsDefault `toml:"default,omitempty"`
-}
-
-type SystemsDefault struct {
-	System     string `toml:"system"`
-	Launcher   string `toml:"launcher,omitempty"`
-	BeforeExit string `toml:"before_exit,omitempty"`
-}
-
-type Launchers struct {
-	IndexRoot   []string `toml:"index_root,omitempty,multiline"`
-	AllowFile   []string `toml:"allow_file,omitempty,multiline"`
-	allowFileRe []*regexp.Regexp
-	Default     []LaunchersDefault `toml:"default,omitempty"`
-	Custom      []LaunchersCustom  `toml:"custom,omitempty"`
-}
-
-type LaunchersDefault struct {
-	Launcher   string `toml:"launcher"`
-	InstallDir string `toml:"install_dir,omitempty"`
-}
-
-type LaunchersCustom struct {
-	ID        string   `toml:"id"`
-	System    string   `toml:"system"`
-	MediaDirs []string `toml:"media_dirs"`
-	FileExts  []string `toml:"file_exts"`
-	Execute   string   `toml:"execute"`
 }
 
 type ZapScript struct {
@@ -109,20 +54,10 @@ type Service struct {
 	allowRunRe []*regexp.Regexp
 }
 
-type MappingsEntry struct {
-	TokenKey     string `toml:"token_key,omitempty"`
-	MatchPattern string `toml:"match_pattern"`
-	ZapScript    string `toml:"zapscript"`
-}
-
-type Mappings struct {
-	Entry []MappingsEntry `toml:"entry,omitempty"`
-}
-
-type Groovy struct {
-	GmcProxyEnabled        bool   `toml:"gmc_proxy_enabled"`
-	GmcProxyPort           int    `toml:"gmc_proxy_port"`
-	GmcProxyBeaconInterval string `toml:"gmc_proxy_beacon_interval"`
+type CredentialEntry struct {
+	Username string `toml:"username"`
+	Password string `toml:"password"`
+	Bearer   string `toml:"bearer"`
 }
 
 var BaseDefaults = Values{
@@ -147,15 +82,16 @@ var BaseDefaults = Values{
 }
 
 type Instance struct {
-	mu      sync.RWMutex
-	appPath string
-	cfgPath string
-	vals    Values
+	mu       sync.RWMutex
+	appPath  string
+	cfgPath  string
+	authPath string
+	vals     Values
 }
 
 func NewConfig(configDir string, defaults Values) (*Instance, error) {
 	cfgPath := os.Getenv(CfgEnv)
-	log.Info().Msgf("env config path: %s", cfgPath)
+	log.Debug().Msgf("env config path: %s", cfgPath)
 
 	if cfgPath == "" {
 		cfgPath = filepath.Join(configDir, CfgFile)
@@ -181,6 +117,8 @@ func NewConfig(configDir string, defaults Values) (*Instance, error) {
 			return nil, err
 		}
 	}
+
+	cfg.authPath = filepath.Join(filepath.Dir(cfgPath), AuthFile)
 
 	err := cfg.Load()
 	if err != nil {
@@ -224,6 +162,23 @@ func (c *Instance) Load() error {
 
 	c.vals = newVals
 
+	// load auth file
+	if _, err := os.Stat(c.authPath); err == nil {
+		log.Info().Msg("loading auth file")
+		authData, err := os.ReadFile(c.authPath)
+		if err != nil {
+			return err
+		}
+
+		var authVals Values
+		err = toml.Unmarshal(authData, &authVals)
+		if err != nil {
+			return err
+		}
+
+		c.vals.Auth = authVals.Auth
+	}
+
 	// prepare allow files regexes
 	c.vals.Launchers.allowFileRe = make([]*regexp.Regexp, len(c.vals.Launchers.AllowFile))
 	for i, allowFile := range c.vals.Launchers.AllowFile {
@@ -265,8 +220,6 @@ func (c *Instance) Load() error {
 		}
 		c.vals.Service.allowRunRe[i] = re
 	}
-
-	log.Info().Any("config", c.vals).Msg("loaded config")
 
 	return nil
 }
@@ -334,88 +287,6 @@ func (c *Instance) SetDebugLogging(enabled bool) {
 	}
 }
 
-func (c *Instance) ReadersScan() ReadersScan {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Readers.Scan
-}
-
-func (c *Instance) IsHoldModeIgnoredSystem(systemID string) bool {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	var blocklist []string
-	for _, v := range c.vals.Readers.Scan.IgnoreSystem {
-		blocklist = append(blocklist, strings.ToLower(v))
-	}
-	return slices.Contains(blocklist, strings.ToLower(systemID))
-}
-
-func (c *Instance) TapModeEnabled() bool {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	if c.vals.Readers.Scan.Mode == ScanModeTap {
-		return true
-	} else if c.vals.Readers.Scan.Mode == "" {
-		return true
-	} else {
-		return false
-	}
-}
-
-func (c *Instance) HoldModeEnabled() bool {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Readers.Scan.Mode == ScanModeHold
-}
-
-func (c *Instance) SetScanMode(mode string) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.vals.Readers.Scan.Mode = mode
-}
-
-func (c *Instance) SetScanExitDelay(exitDelay float32) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.vals.Readers.Scan.ExitDelay = exitDelay
-}
-
-func (c *Instance) SetScanIgnoreSystem(ignoreSystem []string) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.vals.Readers.Scan.IgnoreSystem = ignoreSystem
-}
-
-func (c *Instance) Readers() Readers {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Readers
-}
-
-func (c *Instance) AutoDetect() bool {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Readers.AutoDetect
-}
-
-func (c *Instance) SetAutoDetect(enabled bool) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.vals.Readers.AutoDetect = enabled
-}
-
-func (c *Instance) SetReaderConnections(rcs []ReadersConnect) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.vals.Readers.Connect = rcs
-}
-
-func (c *Instance) SystemDefaults() []SystemsDefault {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Systems.Default
-}
-
 func (c *Instance) IndexRoots() []string {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
@@ -437,12 +308,6 @@ func checkAllow(allow []string, allowRe []*regexp.Regexp, s string) bool {
 	return false
 }
 
-func (c *Instance) IsLauncherFileAllowed(s string) bool {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return checkAllow(c.vals.Launchers.AllowFile, c.vals.Launchers.allowFileRe, s)
-}
-
 func (c *Instance) ApiPort() int {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
@@ -455,192 +320,37 @@ func (c *Instance) IsExecuteAllowed(s string) bool {
 	return checkAllow(c.vals.ZapScript.AllowExecute, c.vals.ZapScript.allowExecuteRe, s)
 }
 
-func (c *Instance) LoadMappings(mappingsDir string) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	_, err := os.Stat(mappingsDir)
-	if err != nil {
-		return err
-	}
-
-	var mapFiles []string
-
-	err = filepath.WalkDir(
-		mappingsDir,
-		func(path string, d os.DirEntry, err error) error {
-			if err != nil {
-				return err
-			}
-
-			if d.IsDir() {
-				return nil
-			}
-
-			if strings.ToLower(filepath.Ext(d.Name())) != ".toml" {
-				return nil
-			}
-
-			mapFiles = append(mapFiles, path)
-
-			return nil
-		},
-	)
-	if err != nil {
-		return err
-	}
-	log.Info().Msgf("found %d mapping files", len(mapFiles))
-
-	filesCounts := 0
-	mappingsCount := 0
-
-	for _, mapPath := range mapFiles {
-		log.Debug().Msgf("loading mapping file: %s", mapPath)
-
-		data, err := os.ReadFile(mapPath)
-		if err != nil {
-			log.Error().Msgf("error reading mapping file: %s", mapPath)
-			continue
-		}
-
-		var newVals Values
-		err = toml.Unmarshal(data, &newVals)
-		if err != nil {
-			log.Error().Msgf("error parsing mapping file: %s", mapPath)
-			continue
-		}
-
-		c.vals.Mappings.Entry = append(c.vals.Mappings.Entry, newVals.Mappings.Entry...)
-
-		filesCounts++
-		mappingsCount += len(newVals.Mappings.Entry)
-	}
-
-	log.Info().Msgf("loaded %d mapping files, %d mappings", filesCounts, mappingsCount)
-
-	return nil
-}
-
-func (c *Instance) Mappings() []MappingsEntry {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Mappings.Entry
-}
-
 func (c *Instance) IsRunAllowed(s string) bool {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
 	return checkAllow(c.vals.Service.AllowRun, c.vals.Service.allowRunRe, s)
 }
 
-func (c *Instance) LookupSystemDefaults(systemId string) (SystemsDefault, bool) {
+func (c *Instance) LookupAuth(reqURL *url.URL) *CredentialEntry {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
-	for _, defaultSystem := range c.vals.Systems.Default {
-		if strings.EqualFold(defaultSystem.System, systemId) {
-			return defaultSystem, true
-		}
-	}
-	return SystemsDefault{}, false
-}
 
-func (c *Instance) LookupLauncherDefaults(launcherId string) (LaunchersDefault, bool) {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	for _, defaultLauncher := range c.vals.Launchers.Default {
-		if strings.EqualFold(defaultLauncher.Launcher, launcherId) {
-			return defaultLauncher, true
-		}
-	}
-	return LaunchersDefault{}, false
-}
-
-func (c *Instance) GmcProxyEnabled() bool {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Groovy.GmcProxyEnabled
-}
-
-func (c *Instance) GmcProxyPort() int {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Groovy.GmcProxyPort
-}
-
-func (c *Instance) GmcProxyBeaconInterval() string {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Groovy.GmcProxyBeaconInterval
-}
-
-func (c *Instance) LoadCustomLaunchers(launchersDir string) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	_, err := os.Stat(launchersDir)
-	if err != nil {
-		return err
-	}
-
-	var launcherFiles []string
-
-	err = filepath.WalkDir(
-		launchersDir,
-		func(path string, d os.DirEntry, err error) error {
-			if err != nil {
-				return err
-			}
-
-			if d.IsDir() {
-				return nil
-			}
-
-			if strings.ToLower(filepath.Ext(d.Name())) != ".toml" {
-				return nil
-			}
-
-			launcherFiles = append(launcherFiles, path)
-
-			return nil
-		},
-	)
-	if err != nil {
-		return err
-	}
-	log.Info().Msgf("found %d custom launcher files", len(launcherFiles))
-
-	filesCounts := 0
-	launchersCount := 0
-
-	for _, launcherPath := range launcherFiles {
-		log.Debug().Msgf("loading custom launcher: %s", launcherPath)
-
-		data, err := os.ReadFile(launcherPath)
+	for k, v := range c.vals.Auth {
+		defURL, err := url.Parse(k)
 		if err != nil {
-			log.Error().Msgf("error reading custom launcher: %s", launcherPath)
+			log.Error().Msgf("invalid auth url: %s", k)
 			continue
 		}
 
-		var newVals Values
-		err = toml.Unmarshal(data, &newVals)
-		if err != nil {
-			log.Error().Msgf("error parsing custom launcher: %s", launcherPath)
+		if !strings.EqualFold(defURL.Scheme, reqURL.Scheme) {
 			continue
 		}
 
-		c.vals.Launchers.Custom = append(c.vals.Launchers.Custom, newVals.Launchers.Custom...)
+		if !strings.EqualFold(defURL.Host, reqURL.Host) {
+			continue
+		}
 
-		filesCounts++
-		launchersCount += len(newVals.Launchers.Custom)
+		if !strings.HasPrefix(reqURL.Path, defURL.Path) {
+			continue
+		}
+
+		return &v
 	}
-
-	log.Info().Msgf("loaded %d files, %d custom launchers", filesCounts, launchersCount)
 
 	return nil
-}
-
-func (c *Instance) CustomLaunchers() []LaunchersCustom {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	return c.vals.Launchers.Custom
 }
